Refresh members of review team
Fixes #5

Logic of this method may be a little obtuse. Here's what it does.

If the current `@members` array is

`[B A C D]`

And the new list of team members is

`[A B D E]`

Then you want your resulting `@members` array to be

`[B A D E]`

In this you

- preserve ordering, so that PRs are distributed fairly
- Drop the member that's now gone
- Add the new member to the end of the collection

The implementation

```ruby
@members = (members | get_members) & get_members
```

Uses Ruby's set operators

```ruby
(members | get_members)
```

Returns the intersection of our cached members and the members on
GitHub. In our example, that would be

```
[B A C D E]
```

This is then intersected with the current members on GitHub, resulting
in

```
[B A D E]
```

Refreshing the team list every 30 seconds is ridiculous, but does no
harm.