Remove most_recent_check
This was introduced to reduce checks on Repos that have not changed. But
it also introduces a bug.

- Person creates a PR
- PRRR assigns a reviewer
- Someone removes  the Reviewer
- PRRR never notices this because the repo's "pushed_at" time remains in
the past.

So, I'm taking the check out. PRRR will just check all repos. This slows
it down a bit. Or more than a bit if you have it checking a lot of repos